Start building base class IAPHandler
